created locators.json
