+++ applied_jobs.yaml
@@ -1,4 +1,6 @@
-https://jobs.jobvite.com/careers/amendllc/job/oQNRufwI: Failed
+https://jobs.jobvite.com/careers/amendllc/job/oQNRufwI: Already Submitted
+https://jobs.jobvite.com/careers/confluent/job/oC4zvfwu: Successfully Applied
 https://jobs.jobvite.com/careers/egnyte/job/o2Cgvfw9: Already Submitted
-https://jobs.jobvite.com/careers/egnyte/job/oIogvfwB: Already Submitted
+https://jobs.jobvite.com/careers/egnyte/job/oIogvfwB: Failed
 https://jobs.jobvite.com/splunk-careers/job/oQQAvfwv: Already Submitted
+https://jobs.jobvite.com/windriver/job/o5yGvfwy: Already Submitted
