+++ job_application_config.yaml
@@ -10,9 +10,9 @@ city: "New York"
 state: "New York"
 zip: "10001"
 country: "United States"
-how did you hear about this job: "LinkedIn Job Posting"
+job posting: "LinkedIn Job Posting"
 degree: "Bachelor"
-college/university: "Acibadem University"
+major degree: "Acibadem University"
 work experience: "5-7 years"
 referred: "No"
 salary: "$100,000 - $120,000"
