+++ jobs/linkedin_jobs.csv
@@ -1,10 +1,14 @@
 platform,url, company, jobid
 
+jobvite,/windriver/job/o5yGvfwy,windriver,o5yGvfwy
+jobvite,/splunk-careers/job/oQQAvfwv,splunk,oQQAvfwv
+jobvite,/careers/confluent/job/oC4zvfwu,confluent,oC4zvfwu
+
+
 jobvite,/careers/amendllc/job/oQNRufwI,amendllc,oQNRufwI
 
-jobvite,/splunk-careers/job/oQQAvfwv,splunk,oQQAvfwv
 
-jobvite,/careers/egnyte/job/oIogvfwB,egnyte,oIogvfwB
+
 jobvite,/careers/egnyte/job/o2Cgvfw9,egnyte,o2Cgvfw9
